Allow customizing release workflow
