+++ project-keeper/src/main/resources/templates/.github/workflows/release.yml
@@ -1,5 +1,3 @@
-# Generated by Project Keeper
-# https://github.com/exasol/project-keeper/blob/main/project-keeper/src/main/resources/templates/.github/workflows/release.yml
 name: Release
 on:
   # [impl->dsn~release-workflow.triggers~1]
